requirements.txt
asgiref==3.3.1
pytz==2021.1
sqlparse==0.4.1
Django==3.1.7
ipdb==0.13.4
djangorestframework==3.11.0
